refactor: persistence 하위의 entity 패키지 제거, persistence에 -Entity 클래스 평탄화
