+++ src/main/java/kr/hhplus/be/commerce/cash/application/CashReadBalanceQueryManager.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.commerce.cash.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import kr.hhplus.be.commerce.cash.persistence.CashEntity;
 import kr.hhplus.be.commerce.cash.persistence.CashJpaRepository;
-import kr.hhplus.be.commerce.cash.persistence.entity.CashEntity;
 import kr.hhplus.be.commerce.cash.presentation.response.CashDetailResponse;
 import kr.hhplus.be.commerce.global.exception.CommerceCode;
 import kr.hhplus.be.commerce.global.exception.CommerceException;
