+++ src/main/java/kr/hhplus/be/commerce/cash/persistence/CashHistoryJpaRepository.java
@@ -2,7 +2,5 @@ package kr.hhplus.be.commerce.cash.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kr.hhplus.be.commerce.cash.persistence.entity.CashHistoryEntity;
-
 public interface CashHistoryJpaRepository extends JpaRepository<CashHistoryEntity, Long> {
 }
