+++ src/main/java/kr/hhplus/be/commerce/cash/persistence/CashHistoryRepository.java
@@ -1,7 +1,5 @@
 package kr.hhplus.be.commerce.cash.persistence;
 
-import kr.hhplus.be.commerce.cash.persistence.entity.CashHistoryEntity;
-
 public interface CashHistoryRepository {
 	CashHistoryEntity save(CashHistoryEntity cashHistoryEntity);
 }
