+++ src/main/java/kr/hhplus/be/commerce/cash/persistence/CashRepository.java
@@ -2,8 +2,6 @@ package kr.hhplus.be.commerce.cash.persistence;
 
 import java.util.Optional;
 
-import kr.hhplus.be.commerce.cash.persistence.entity.CashEntity;
-
 public interface CashRepository {
 	Optional<CashEntity> findByUserId(Long userId);
