+++ src/main/java/kr/hhplus/be/commerce/coupon/application/UserCouponIssueProcessor.java
@@ -6,10 +6,10 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import kr.hhplus.be.commerce.coupon.persistence.CouponEntity;
 import kr.hhplus.be.commerce.coupon.persistence.CouponJpaRepository;
+import kr.hhplus.be.commerce.coupon.persistence.UserCouponEntity;
 import kr.hhplus.be.commerce.coupon.persistence.UserCouponRepository;
-import kr.hhplus.be.commerce.coupon.persistence.entity.CouponEntity;
-import kr.hhplus.be.commerce.coupon.persistence.entity.UserCouponEntity;
 import kr.hhplus.be.commerce.global.exception.CommerceCode;
 import kr.hhplus.be.commerce.global.exception.CommerceException;
 import lombok.RequiredArgsConstructor;
