+++ src/main/java/kr/hhplus/be/commerce/coupon/persistence/UserCouponJpaRepository.java
@@ -2,8 +2,6 @@ package kr.hhplus.be.commerce.coupon.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kr.hhplus.be.commerce.coupon.persistence.entity.UserCouponEntity;
-
 public interface UserCouponJpaRepository extends JpaRepository<UserCouponEntity, Long> {
 	boolean existsByUserIdAndCouponId(Long userId, Long couponId);
 }
