+++ src/main/java/kr/hhplus/be/commerce/coupon/presentation/response/UserCouponSummaryResponse.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.commerce.coupon.presentation.response;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-import kr.hhplus.be.commerce.coupon.persistence.entity.enums.CouponDiscountType;
-import kr.hhplus.be.commerce.coupon.persistence.entity.enums.UserCouponStatus;
+import kr.hhplus.be.commerce.coupon.persistence.enums.CouponDiscountType;
+import kr.hhplus.be.commerce.coupon.persistence.enums.UserCouponStatus;
 
 public record UserCouponSummaryResponse(
 	Long id,
