+++ src/main/java/kr/hhplus/be/commerce/order/infrastructure/persistence/OrderLineJpaRepository.java
@@ -4,8 +4,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kr.hhplus.be.commerce.order.infrastructure.persistence.entity.OrderLineEntity;
-
 public interface OrderLineJpaRepository extends JpaRepository<OrderLineEntity, Long> {
 	List<OrderLineEntity> findAllByOrderId(Long orderId);
 }
