+++ src/main/java/kr/hhplus/be/commerce/payment/infrastructure/persistence/PaymentJpaRepository.java
@@ -2,7 +2,5 @@ package kr.hhplus.be.commerce.payment.infrastructure.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kr.hhplus.be.commerce.payment.infrastructure.persistence.entity.PaymentEntity;
-
 public interface PaymentJpaRepository extends JpaRepository<PaymentEntity, Long> {
 }
